REFACTOR: Move project delete and basic client routes to blueprints
- Added project delete route to projects blueprint
- Created comprehensive clients blueprint with:
  - list_clients, create_client, view_client routes
  - Proper access control and activity logging
  - Contact association functionality
- Removed 4 routes from app.py (80 → 76 routes remaining)
- Maintained all existing functionality and template compatibility
