+++ app.py
@@ -341,48 +341,7 @@ def generate_access_code_route():
 
 
 
-@app.route('/projects/<int:id>/delete', methods=['POST'])
-def delete_project(id):
-    project = Project.query.get_or_404(id)
-    
-    # Check access permission
-    if project.firm_id != session['firm_id']:
-        return jsonify({'success': False, 'message': 'Access denied'}), 403
-    
-    try:
-        project_name = project.name
-        client_name = project.client_name
-        
-        # Get count of associated tasks for confirmation
-        task_count = Task.query.filter_by(project_id=id).count()
-        
-        # Create activity log before deletion
-        create_activity_log(f'Project "{project_name}" for {client_name} deleted (with {task_count} tasks)', session['user_id'])
-        
-        # Delete all associated tasks first (cascade deletion)
-        tasks = Task.query.filter_by(project_id=id).all()
-        for task in tasks:
-            # Delete task comments first
-            TaskComment.query.filter_by(task_id=task.id).delete()
-            # Delete the task
-            db.session.delete(task)
-        
-        # Delete project attachments if any
-        Attachment.query.filter_by(project_id=id).delete()
-        
-        # Delete the project itself
-        db.session.delete(project)
-        db.session.commit()
-        
-        return jsonify({
-            'success': True, 
-            'message': f'Project "{project_name}" and {task_count} associated tasks deleted successfully',
-            'redirect': '/projects'
-        })
-    
-    except Exception as e:
-        db.session.rollback()
-        return jsonify({'success': False, 'message': f'Error deleting project: {str(e)}'}), 500
+# Project delete route moved to projects blueprint
 
 
 # Task edit route moved to tasks blueprint
@@ -482,53 +441,7 @@ def create_user():
     
     return render_template('admin/create_user.html')
 
-@app.route('/clients')
-def clients():
-    firm_id = session['firm_id']
-    clients = Client.query.filter_by(firm_id=firm_id).order_by(Client.name.asc()).all()
-    return render_template('clients/clients.html', clients=clients)
-
-@app.route('/clients/create', methods=['GET', 'POST'])
-def create_client():
-    if request.method == 'POST':
-        firm_id = session['firm_id']
-        
-        client = Client(
-            name=request.form.get('name'),
-            email=request.form.get('email'),
-            phone=request.form.get('phone'),
-            address=request.form.get('address'),
-            contact_person=request.form.get('contact_person'),
-            entity_type=request.form.get('entity_type', 'Individual'),
-            tax_id=request.form.get('tax_id'),
-            notes=request.form.get('notes'),
-            firm_id=firm_id
-        )
-        db.session.add(client)
-        db.session.commit()
-        
-        # Activity log
-        create_activity_log(f'Client "{client.name}" created', session['user_id'])
-        
-        flash('Client created successfully!', 'success')
-        return redirect(url_for('clients'))
-    
-    return render_template('clients/create_client.html')
-
-@app.route('/clients/<int:id>')
-def view_client(id):
-    client = Client.query.get_or_404(id)
-    if client.firm_id != session['firm_id']:
-        flash('Access denied', 'error')
-        return redirect(url_for('clients'))
-    
-    projects = Project.query.filter_by(client_id=id).order_by(Project.created_at.desc()).all()
-    
-    # Get available contacts (not already associated with this client)
-    associated_contact_ids = db.session.query(ClientContact.contact_id).filter_by(client_id=id).subquery()
-    available_contacts = Contact.query.filter(~Contact.id.in_(associated_contact_ids)).all()
-    
-    return render_template('clients/view_client.html', client=client, projects=projects, available_contacts=available_contacts)
+# Client routes moved to clients blueprint
 
 @app.route('/calendar')
 def calendar_view():
